Use uintptr_t type for ThreadID
For Arduino target code there is no difference between int and pointer sizes, e.g.
$HOME/.arduino15/packages/arduino/tools/avr-gcc/7.3.0-atmel3.6.1-arduino4/avr/include/stdint.h
   typedef uint16_t uintptr_t;

but it matters when compiling the code for unit test on a 64bit linux host computer:

.../arduino/sketchbook/libraries/ArduinoThread/Thread.cpp: In constructor ‘Thread::Thread(void (*)(), long unsigned int)’:
.../arduino/sketchbook/libraries/ArduinoThread/Thread.cpp:9:18: error: cast from ‘Thread*’ to ‘int’ loses precision [-fpermissive]
  ThreadID = (int)this;
                  ^~~~

+++ Thread.h
@@ -61,7 +61,7 @@ public:
 	bool enabled;
 
 	// ID of the Thread (initialized from memory adr.)
-	int ThreadID;
+	uintptr_t ThreadID;
 
 	#ifdef USE_THREAD_NAMES
 		// Thread Name (used for better UI).
